move clean_up_heatmap and integrate it into quantify_motifs
